qsdqsd

+++ src/main/java/be/delomid/oneapp/mschat/mschat/dto/BuildingDto.java
@@ -3,6 +3,7 @@ package be.delomid.oneapp.mschat.mschat.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -16,10 +17,10 @@ public class BuildingDto {
     private Integer yearOfConstruction;
     private Integer numberOfFloors;
     private String buildingState;
-    private Double facadeWidth;
-    private Double landArea;
-    private Double landWidth;
-    private Double builtArea;
+    private BigDecimal facadeWidth;
+    private BigDecimal  landArea;
+    private BigDecimal  landWidth;
+    private BigDecimal  builtArea;
     private Boolean hasElevator;
     private Boolean hasHandicapAccess;
     private Boolean hasPool;
